requirements.txt
python-dotenv
pandas
numpy
statsmodels
streamlit
SQLAlchemy
pymysql
matplotlib